Fix : Chat
